cli/pyproject.toml
[project]
name = "nao-core"
version = "0.0.36"
description = "nao Core is your analytics context builder with the best chat interface."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "nao Labs" }]
keywords = ["analytics", "chat", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "cyclopts>=4.4.4",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "ibis-framework[bigquery,duckdb,databricks,mssql,snowflake,postgres,athena]>=9.0.0",
    "openai>=1.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "cryptography>=46.0.3",
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "jinja2>=3.1.0",
    "apscheduler>=3.10.0",
    "anthropic>=0.76.0",
    "notion-client>=2.7.0",
    "notion2md>=2.9.0",
    "questionary>=2.1.0",
    "posthog>=7.8.0",
    "mistralai>=1.11.1",
    "google-genai>=1.61.0",
    "sshtunnel>=0.4.0",
    "snowflake-connector-python[secure-local-storage]>=4.2.0",
    "atlassian-python-api>=3.41.0",
]

[project.optional-dependencies]
dev = ["pytest-cov"]

[project.scripts]
nao = "nao_core.main:main"

[project.urls]
Homepage = "https://getnao.io"
Repository = "https://github.com/naolabs/chat"

[dependency-groups]
dev = ["ruff>=0.8.0", "ty>=0.0.9"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
# Include the bin directory with server binary and public assets
artifacts = ["nao_core/bin/**"]

[tool.hatch.build.targets.wheel]
packages = ["nao_core"]
exclude = ["**/example/**"]

[tool.hatch.build.targets.sdist]
include = [
    "nao_core/**",
    "README.md",
    "LICENSE",
]
exclude = ["**/example/**"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
